tests: rework tests to use JSON requests
This is still not *as* E2E as, say, mocking an actual LSP client and
then ensuring everything is handled properly, but this is basically as
good as it gets without massively overcomplicating things. This is also
the recommended way to test things based on the `tower_lsp`
documentation, and the way the `sway-lsp` does it (which provides a few
good examples).