feat: standalone formatter (`ts_query_ls format`)
Allows the language server to act as a standalone formatter by passing
the `format` argument. It will then take a list of directories to
format, and a formatting "mode" (whether to just check the files or
actually format them). Formatting is parallelized, and quite quick for
large query repositories like `nvim-treesitter`.