feat!: diagnostics for impossible patterns (#115)
This will spawn a new query with `Query::new` for every pattern
definition in the file. This way we can get more than one impossible
pattern diagnostic per file. This creates a lot of overhead, especially
for bugged queries that max out the iteration count before terminating
analysis (with no helpful results). Thus, we cache the results of the
structure analysis by query text and language name. So additional work
is only performed when a pattern changes and its analysis is *not* found
in the cache.

This commit also removes code action/command support, which is why it is marked as breaking.