test: overhaul initialization logic
This is a large refactor that separates the document population logic
from the language population logic. This allows us to craft much more
granular tests, such as ones that check for diagnostics when no
languages are found. It also paves the way to using real language
objects in tests, to provide full end-to-end coverage. Additionally, it
allows tests to be fully workspace-aware, allowing coverage for
workspace methods (e.g. workspace symbols), and for diagnostics which
require importing files from the workspace.

It also removes the custom test document population logic with actual
calls to `textDocument/didOpen`, giving full coverage for that
notification.